Move `ProjectAvatar` to its onw file
